Reduce the visibility of GVRComponent.getOwnerObjecct() from public to protected
getOwnerObject() was used to set CameraRig transform, adding camera-frame objects (pickers,
head trackers, moved objects, etc.), and set orientation data from sensors. Because
implementations of these functionalities are subject to change, it is risky to expose the
owner object to users.

Recently, new APIs have been added to CameraRig to directly perform these functions, such as
getTransform(), addChildObject, removeChildObject, etc. So it is less needed to keep getOwnerObject()
public; further, when is is public, there are multiple ways to do common things and can lead to
confusion and maintainability issues.

The proposed solution is to
   1) Make GVRComponent.getOwnerObject() protected.
   2) Add explicit API to perform things for CameraRig, such as attachToParent() and
      detachFromParent().
   3) An exception is to make GVREyePointeeHolder.getOwnerObject() public to support current functionalites.
      Refactoring of this class is left for others if seen appropriate.