GVREventManager: Always check method name and parameter types in sendEvent
If user code specifies a wrong method name, wrong number of parameters,
or a parameter of a wrong type, throw RuntimeException, instead of ignoring
it silently.

This was implemented in the initial version, but was skipped after listeners
are supported. Now a slightly different version is added that works in the
listener case as well as the subtyping case.